Add tests for the Hover Provider and Message Router
This commit adds tests for the Hover Provider and the Message Router.  This
commit also adds helper Mock objects for the JSON RPC handler and Pupept
Resource objects which are needed in unit testing.  The `--no-preload` option
is added in spec setup to speed up tests which do not require the full Puppet
initialization, typically these are the unit tests.