Update to support Puppet 5 AST
In Puppet 5, the AST changed dramatically mainly due to removing the
parent pointer 'eContainer' from the nodes. Also many methods were
renamed and their implementation changed slightly. 'eContents' became
'_pcore_contents' and eAllContents became '_pcore_all_contents' with
some subtlties that allow you to record the traversal path to each node.

This changes the parser_helper to return both the node under the cursor
and the path to that node. For Puppet < 5 it uses 'eContainer' to go up
the AST, and for Puppet 5 it records the paths using
_pcore_all_contents.