Fix completion provider with Puppet 5.2.0
Puppet issue 7668 (https://tickets.puppetlabs.com/browse/PUP-7668) was fixed in
the puppet gem 5.2.0, which then caused parsing to work correctly.  However this
meant that root document checks were returning a BlockExpression object instead
of nil.  This commit changes the parsing so that the BlockExpression class as it
is a container for other objects and can not be auto-completed in of itself.

This commit also adds testing of Puppet gem 5.1.0 as this version of Puppe does
not contain the PUP-7668 fix and we must ensure that all puppet versions have
the same autocomplete behaviour.