(maint) Enable simple tcp server to exit when sent sighup
Previously the Simple TCP server would not exit when a SIGHUP (Ctrl-C) signal
was sent.  This commit changes the stop_services handler so that it is aware
when it is invoked from a TRAP as some operations are not allowed in the TRAP
context.  This commit also updates the trap handlers to use the correct
method name.