(GH-231) Make document validation asynchronous
The document validation/diagnostics part of the language server is executed on
every document open or change event. While normally this is ok, when typing fast
on large documents, this can create slowness. Particularly as we add more
features which require longer running diagnostic processes.  This commit uses
a thread safe queue system which will only allow the document to exist in
the queue once.  Safeguards are also added which will ignore validation results
if the document is changed mid-validation.