(GH-236) Add an experimental persistant cache layer
Previously the puppet assets were always loaded from Puppet on every invocation
of the Langauge server.  This incurs a large startup penalty.  This commit adds
the ability to save and load cache results to disk.  The cache results use the
SHA256 of the source files and version information to invalidate cache entries.