(maint) Enforce puppet loading order for facts and types
Previously facts and puppet types were loaded asynchronously which was causing
an unseen issue whereby some types (e.g. Service, Package) were not seen as
loaded.  This was due to gem based facter loading the these types before the
type loader.  This would not normally be in an issue in Puppet Agent as
native facter would not share the same ruby environment.  However in development
, or soon to add, PDK workflows, a gem based facter would be used.

This commit changes the load order to synchronously load the types first and
then load the facts, classes and functions asynchronously.