(GH-236) Implement a fake object layer for Puppet assets
Previously the Language Server would query Puppet directly for lists of objects
such as types or functions.  However this meant for simple query operations the
entire object had to be parsed and loaded.  his commit introduces an object
system that can be populated from Puppet and then be consumed by the Language
Server (faux_objects.rb).