(maint) Stop invalid debug logs from crashing services
Previously if an invalid logfile location was specified e.g. C:\ or a missing
parent directory, the service (Language or Debug Server) would error.  This is
undesired as a bad log file should not cause a catastrophic failure.  This
commit modifies the logging so that an error occurs when the log file is
created it, the error is sent to STDERR, and service continues but with logging
disabled.

This commit also flushes the log file on each write if needed.  This is required
as OS buffering makes real time debugging very difficult.