(maint) Initialize the language server synchronously in rspec
Previously some unit tests were failing when run on their own, but succeeded
when run all together.  This was due to the puppet initialization taking place
in another thread which caused a small race condition, only preset in rspec.
This commit changes the load behavior to be sychronous as possible.  Note that
class/type/function loading is still async, but the helper method
`wait_for_puppet_loading` can be called to wait for loading to complete.