(GH-14) Ignore environmentpath puppet setting if it does not exist
Previously the language server would crash if the Environment Path did not exist.
This can happen when pluginsync syncs all the files into the lib directory, or a
newly created puppet agent.  This commit changes the load behaviour to instead
check if the directory exists and just ignore it if it does not.  This commit
also adds a test for this scenario.