(GH-10) Disable the file cache if temp directory doesn't exist
Previously the persistent file cache would crash if the temporary directory
could not be detected, or did not exist.  This commit changes the detection to
be cross platform supported and adds error detection if the cache directory
could not be created.  If the the cache directory is unable to be created all
calls to the persistent file cache then emulate as if the cache is empty thereby
still allowing the language server to operate.  This commit also adds tests for
this scenario.