(GH-26) Refactor workspace detection
Previously the --local-workspace option was mostly ignored in the puppetfile
validation.  However this has caused problems when validating inside control
repositories which do not have a metadata.json file (instead it's Puppetfile).
This commit refactors the workspace detection out of the puppetfile validation
and moves it into the Document Store.

This commit also adds a simple timed based caching mechanism to reduce File IO
as it can be exercised vigourously when users are typing and validation is
occuring very often.  An arbitrary value of 60 seconds was chosen the TTL of the
cache.  This means that after 60 seconds the workspace will be re-evaluated to
make sure the user is still in a module (metadata.json) or control repo
(Puppetfile).