(GH-31) Use canonical names for line based breakpoints
Previously VS Code was sending filenames in the API calls which always matched
the file system, however this is no longer the case.  This exposed a bug in the
debug adapter which depended on this behaviour. This commit modifies the line
based breakpoints to use a canonical filename.  However due to debug server
never knowing what naming scheme is coming from the editor we had to resort to
a naive method of just setting it to all lowercase.  This _may_ cause an issue
on case sensitive file systems however Puppet style guides prohibit manifest
names only changing by case.