(GH-40) Initial Puppet Language Sidecar
Previously all of the Puppet helpers ran inside of a single instance of Puppet,
however this causes issues because classes and types cannot be unloaded in a
running instance. So instead, a sidecar executable is created which will run
out of process, in a fresh Puppet configuration, to enable enumerating types,
classes and functions etc.

This commit creates the basic skeleton of this sidecar process.