(GH-40) Add Sidecar capabilities to enumerate classes, functions and types
This commit adds the ability for the sidecar process to enumerate classes,
functions and types, either in the default configuration (i.e. like a real
puppet agent) or only within the confines of a workspace directory (e.g. when
editing a module or control-repo). Currently this only supports querying a
puppet module workspace.

This commit also adds the ability to pass in puppet options (e.g. --vardir) into
the puppet configuration for the sidecar.

This commit also uses a filesystem based caching mechanism to speed up the
default enumerations.

The majority of this code was extracted from the existing PuppetHelper in the
Language Server and Puppet Monkey Patches.