(GH-40) Use sidecar to retrieve puppet information
Previously a sidecar process was introduced to query puppet for information,
however the Language Server was yet to take advantage of it.  This commit
modifies the Puppet Helper in the Language Server to use the sidecar instead.

This commit also uses a queueing mechanism to throttle the sidecar and ensure
that if there are duplicate sidecar requests, only the most recent is processed.
This ensures that that CPU and memory use is capped, but still maintains a good
repsonse time and user experience.