(GH-40) Add sidecar serialisation protocol
The Language Server and the sidecar need an agreed upon protocol to communicate
properly.  The agreed upon encoding will be UTF8 JSON strings.  This commit adds
the required classes which can be used to serialise and deserialise objects to
and from the sidecar process.  This commit also includes tests to validate the
serialisation process.