(GH-49) Exit STDIO loop if STDIN reaches EOF
Previously if the STDIN pipe was in a bad state, the pipe would appear to be
readable however when it is attempted to be read, it was throwing an EOF.  This
can occur when Javascript closes a child language server process.  This wasn't
caught in development due to the INT trap that is setup globally, so when
developers pressed Ctrl-C in the javascript parent process, the ruby child
process also saw it and terminated correctly.