(GH-54) Fix Type loading for Puppet 6
With the new puppet 6.0 gem, the type loading in the Language Server no longer
worked.  This was due autoloader methods changing their method signature. Given
that this is a major version bump, breaking changes are not unexpected.  This
commit changes:
* The file loader to always pass the environment parameter. Before Puppet 6 this
  was optional, but it is now a required parameter
* The files_to_load method has changed from no parameters to one.  This commit
  modifies the call to first check the arity of the method, i.e. how many
  parameters are needed, and then modifies the file_to_load call appropriately.

  Using arity is better than checking Puppet version as it takes into account
  Puppet 7 and beyond.

Note that no test changes were made and this confirms that the language server
behaviour is the same across Puppet 4,5 and 6.