(GH-54) Fix Function loading for Puppet 6
With the new puppet 6.0 gem, the function loading in the Language Server no
longer worked.  This was due autoloader methods changing their method signature
and the Puppet::Parser::Functions.environment_module(env) no longer working as
expected. Given that this is a major version bump, breaking changes are not
unexpected.  This commit changes:
* The monkey patching for functions now keeps track of which functions have
  been loaded through a simple hash table, and is accessible from
  Puppet::Parser::Functions.monkey_function_list
* The array functions filenames.uniq!.compact! could cause nils being generated.
  By changing to filenames.uniq.compact we can guarantee that an empty array
  would be generated, not nils.
* The file loader to always pass the environment parameter. Before Puppet 6 this
  was optional, but it is now a required parameter

Note that no test changes were made and this confirms that the language server
behaviour is the same across Puppet 4,5 and 6.