(maint) Refactor Travis and Appveyor test matrix
Previously each testing combination was tested in isolation.  However this meant
bundler was being called multiple times for no reasone.  Also while the
Appveyor test runtimes are quite low, because there are so many cells there's a
large delay in the Appveyor suite.  This commit refactors the Travis and
Appveyor files to test as much as possible for each Ruby/Puppet combination.
There is no loss of test coverage, just the ordering is different.