(GH-20) Change Puppetfile to detecting environment conf
Previously Control Repositories would be detected by the existance of a file
called Puppetfile.  However this is not true.  Puppetfile is only used to manage
the modules in a control repo.  Instead we should depend on the existance of the
environment.conf file [1]

Strictly speaking environments don't need an environment.conf however there is
no other choice and the typical roles and profiles pattern always lays down this
file.

[1] https://puppet.com/docs/puppet/latest/environments_creating.html