(GH-88) Add a workspace symbol provider
The object cache contains all of the type, function and class information so it
should be easily be able to be queried and then emit these results as part of a
workspace symbol request (as opposed to document symbol).  This commit adds the
provider and cache methods to support querying all objects in the cache.