(maint) Fix parsing at beginning of a document
Previously the parser helper was silently failing to parse documents
that had the cursor at line 0, char 0. This was due to the
removal based methods trying to remove characters at the beginning
of the document.  This commit modifies the helper to ignore these
parsing attempts if the cursor is at line 0, char 0.