(GH-111) Add puppet-version command line argument
Previously it was not possible to specify which puppet gem version to use via
the command line arguments.  This commit adds a --puppet-version argument to the
language server and also the sidecar.  This commit also passes through the
puppet version from the lanugage server to the sidecar

This commit will also issue a warning message to the language client if the
puppet version could not be satisified.