(GH-113) Rescue errors when running Facter 2.x
Previously the facter helper had no error trapping.  Normally this isn't an
issue on Facter 3.x (Puppet Agent), however the PDK uses Facter 2.x which is
still running in the same ruby session as the Language Server.  This means that
if an error occurs then it could crash the entire language server.  This commit
wraps the facter loading in a rescue block and returns zero facts for
intellisense.