(GH-121) Add documentation for Puppet 4 Functions
Previously the Puppet 4 API functions would be loaded however they would not
have any documentation notes as this was removed in the newer API.  Instead we
need to use Puppet Strings and YARD to generate the documentation.  This commit:

Note that Puppet Strings and YARD are only available in the PDK, not Puppet
Agent.  Once this feature is no longer hidden behind a feature flag, they
can be vendored into Editor Services

* Adds the puppet-strings gem for development purpsoes

* Sets up YARD via Puppet Strings which adds the Puppet Language parsing into
YARD

* Caches the result from YARD to speed up documentation resolution for multiple
items in the same file

* Post processes the function loader and adds the documetation for Puppet 4 API
functions