(GH-121) Add a 'pup4api' feature flag
* This commit adds and validates a pup4api feature flag in the sidecar.  This
feature flag will be used to toggle the loading of Puppet information to use the
newer Puppet 4 API.

* This commit copies the helper and monkey patch files for the pup4api flag in
prepration for the work to modify them to use the pup4api

* This commit also copies and modifies the integration tests to use the pup4api
feature flag and ensure that the behaviour does not change when the flag is
enabled

* Note that the pup4api feature flag is gated on the Puppet Gem version being
v6.0 or higher.  This is due to Puppet 6 having all of the required helpers
needed by the Sidecar, even though the API we will be using was introduced in
Puppet 4.