(GH-121) Use file modification time to invalidate the cache
Previously the file cache would use a SHA256 hash of the file content.  While
this works, it's just as easy to use the file modification time stamp. This
reduces the the file IO to check the cache. It's unlikely that a file would have
different content while the modification timestamp stays the same.

Note that this only occurs when the pup4api feature flag is present.