(GH-118) Fail gracefully when critical gems cannot load
Previously the language server would crash/terminate early if a critical gem
like puppet was unavailable.  This can happen when the ruby environment is not
from the Puppet Agent or PDK.  This commit changes the behaviour of the
Language Server to still execute but it a completely disable fashion:

* Detects a failed gem load for critical gems and sets
  PuppetLanguageServer.active? is to false
* When the Language Server is not active, a different Message Router is used
  which effectively warns the user that the server failed to start and that
  all functions are disabled. The server responds to the client with no
  capabilities and the custom getVersion request, is responded to with
  unknown data

No automated tests were added as this is an edge case.  Manual testing was
performed by changing the call `require 'puppet'` to `require 'puppetxxxxx'`
which is enough to trigger a failure.