(GH-144) Add optional remove_trigger_char to Puppet Parser Helper
Previously the object_under_cursor always assumed it should remove the trigger
character when trying to figure out the object under the cursor. While this
works for things like completion and hover, it does not for signature help. This
commit updates the options hash with a new key called remove_trigger_char which
by default removes the character during detection.  Just like the default
behaviour.  This commit does not require any test changes as behaviour has
not changed.