(GH-144) Add Locator to Puppet Parser Helper response
Previously the object_under_cursor method only returned the AST object and the
path to the object.  However some providers may also require a locator because
Puppet 4 does not expose the Locator on the AST object, only on the Factory.

This commit adds the return struct for the object_under_cursor method to also
return the Locator.

No test changes are required.