(GH-147) Gracefully fail on LoadError when compiling manifests
Previously the sidecar would error completely when a LoadError was encountered.
This could be caused by gem Facter trying to load libraries which are not
present.  StandardError does not trap these.  This commit also traps the
LoadError and allows the sidecar to return a properly formatted response to
Editor Services intead of dumping error text to STDERR.