(GH-55) Introspect Debug Adapter Protocol
Previously the ruby version of Debug Adapter Protocol had to be reverse
engineered from the JSON schema. However it is now available as a node module.
This commit extends the LSP introspection process to now also introspects the
Debug Protocol and generates the ruby file.  Later commits will use the
generated code and remove the older classes.