(GH-55) Refactor the Debug Server to run Puppet in the main thread
Previously the debug session ran in the main thread, and puppet was run in a
child thread. However there was a very strange bug in that Ruby 2.5 was
terminating when calling Enumerator.next (and some other methods).  When Puppet
was run in the main thread, and the debug session a child thread, this issue was
not present.

This commit refactors the Puppet Debug Server to run Puppet in the main thread
and the Debug Session and RPC server in child threads.  This commit also:

* Splits up the PupeptDebugSession module into separate classes which makes
  testing and future maintenance easier
* Minimal changes to tests, to prove that the change in implementation does not
  change behaviour
* Reduce the invasiveness of the monkey patching
* Uses the new DSP::* generated protocol classes and removes the old
  DebugServer::Protocol file.