(maint) Add tests for roundtripping hash serialisation
Previously the Sidecar Protocol had methods to convert to and from hash however
they were never tested to ensure that they could be "round-tripped", that is
converted to a hash and then back to a sidecar object with no loss of data. This
commit adds the test and corrects any behvaiour in the protocol implementation.