(GH-163) Add AggregateMetadata sidecar protocol object
Previously, to get all metadata information, the sidecar would need to be called
multiple times, however, with the puppetstrings featureflag, we can now evaluate
the metadata all at once.

This commit adds a new object called AggregateMetadata which can hold all of the
metadata objects (classes etc.) This commit also adds tests to ensure that it
can be serialised and deserialised correctly.

And finally this commit changes the puppetstrings helper to use the new
aggregate instead of a standard hashtable.