(GH-163) Add aggregate sidecar tasks
Previously the AggregateMetadata sidecar protocol object was added.  This commit
adds a default_aggregate and workspace_aggregate sidecar action which will
gather all of the metadata in one call. These actions will only be valid with
the puppetstrings featureflag.

This commt also adds integration tests for these new actions, with and without
the featureflag set.