(maint) Munge reserved words in LSP introspection
Previously the ruby classes created by the LSP introspection could create
methods which are reserved in ruby, but not JavaScript.  This commit modifies
the introspection tool to munge reserved method names with the suffix "__lsp"
so that they can still be used in ruby but the JSON serialisation and
deserialisation still work for the protocol.