(GH-182) PDK Ruby API
This commit adds the following PDK commands to the language server:

- [x] pdk new class
- [x] pdk new defined type
- [x] pdk new task

The commands `pdk new module`, `pdk validate`, and `pdk test unit` were
not implemented in this work. Each command requires user input or
streaming output to make them useful or responsive, which is too large a
set of work to implement in this commit.