(GH-139) Add more intelligence to completion request
Previously the completion provider blindly moved the cursor one place back which
worked fine if completions were triggered by a character instead of manual.
Now that the LSP has provisions for the context of the completion request we
can intelligently figure out when to remove the trigger character.

This commit updates the mesasge handler to pass through the CompletionContext
object and the completion provider can then determine how to use the context
accordingly.  This commit also updates the tests for this behaviour.