(GH-209) Decouple session_state and connection id
Previously when creating a ClientSessionState object it required a "real"
message handler.  Instead this commit updates the initializer so that you can
pass in exactly what the connection_id is, and fallback to message_handler way
if it's nil.  This makes it easier to test and mock.

This commit also removes any unneeded object creations for the session state in
the test suite.