(GH-213) Gather facts using the Sidecar
Previously the facts where evaluated only within the Editor Service process
however this meant the workspace couldn't be injected.  This commit adds a new
action to the Sidecar process to evaluate the facts. The Sidecar has the
required monkey patches to inject the workspace into the fact search
directories.

This commit also adds tests and fixtures to ensure that custom and external
facts can be resolved.