(GH-245) Use object cache for fact data
Previously the puppet/getFacts request would call out to Facter directly however
this can be very slow, particularly with Facter v2 in the PDK.  This commit
changes the request to instead use the object cache which already has all of the
available facter information.

This commit also moves the method from the PuppetHelper to the FacterHelper
where it would make more sense.  This is tested as part of the acceptance suite.