(maint) Remove vendored gems
This commit removes the vendored gems from the git repository and instead
modifies Rake tasks and documentation to vendor the gems manually. Note that
the build process does vendor the gems prior to compressing into a Zip.

This is to reduce git history churn when adding or updating vendored gems. The
vendored gem content doesn't change so there's no point replicating in this
repo.