(GH-189) Reset PuppetLint configuration for each call
Previously after a call to fix_validation_errors, many validation errors came
in as hints not errors.  This was due to PuppetLint not resetting the
`PuppetLint.configuration.fix` setting between calls.  This commit updates the
validation provider to reset this setting each call.  This commit also adds
tests for this scenario.