(GH-306) Add Puppet Lexer for whitespace and comments
Previously the Puppet Lexer ignored whitespace and comments and did not generate
any token information. This meant it could not be used for code formatting
purposes. This commit adds a new lexer which sub-classes the original and emits
tokens for comments. Later commits can use this lexer for operations like code
folding and as-you-type formatting.

This commit also memoizes the tokens in the document store so documents do not
need to be lexed multiple times for the same version.