(MAINT) - TEMP pin to FFI 1.15.2
This change temporarily pins the ffi 1.15.2 gem and pulls this in to be
used in CI for the editor-services.

This change should be temporarily, to resolve failing CI until this bug
is fixed upstream, and released in the next puppet platform release.
See here https://github.com/puppetlabs/puppet/pull/9112