(CAT-1493) - Fix missing file resource type parameters
Prior to this commit, when using the autocompletion feature of
puppet-editor-services, the file resource type would be missing some
parameters/properties in the dropdown completion list.

It was found that this was due to the way the file resource type was
structured in the puppet source code. The type defintion can be found in
both /lib/puppet/type and /lib/puppet/type/file/.

The way we can work around this is by altering the search glob,
and combining the type defintions into one single file. This file is
then deleted on completion.