(CAT-1595) - Remove diagnostic on textDoucment onDidClose
Prior to this commit, diagnostics (or "problems") in the vscode problems
pane would not be removed on the closing of a file. This is bad
practice, and quickly leads to a polluted problems pane with linting
issues etc showing for all files despite being open or not.

This commit aims to fix this issue, by calling publishDiagnostics in the
vscode API on the close of a text document with an empty array of
diagnostics as per the vscode documentation.

This has been verified as working in local vscode setup.