CAT-1670 - Fix undefined variable json_rpc_handler
Prior to this commit, we would see in the language server logs that we
were referencing an undefined variable `json_rpc_handler`.

This had knock on effects, such as registering the onTypeformatting
provider, meaning this setting would not work when enabled.

We now update the json_rpc_handler send message event to be consistent
with the rest of the language server, and to fix the undefined variable
error.

The dynamically registered providers, like onTypeFormatting, will now
work as expected.